src/main/java/com/cheese/demo/coupon/CouponServiceImpl.java
package com.cheese.demo.coupon;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
public class CouponServiceImpl implements CouponService {

    private final CouponRepository couponRepository;

    @Autowired
    public CouponServiceImpl(CouponRepository couponRepository) {
        this.couponRepository = couponRepository;
    }

    @Override
    @Transactional
    public Coupon create(CouponDto.Creation dto) {
        return couponRepository.save(dto.toEntity());
    }
}